// src/app/customers/models/customer.model.ts
export interface Customer {
    id?: string;
    firstName: string;
    lastName: string;
    age: number;
    title: string;
    other?: string;
    company: string;
    email: string;
    phones: string[];
    address: string;
    address2: string;
    city: string;
    state: string;
    zipCode: string;
}